fixed es6 syntax and modules
