+++ index.html
@@ -97,7 +97,7 @@
         <span class="text-muted">Copyright</span>
       </div>
     </footer>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
     <script
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
       integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
