+++ package.json
@@ -13,5 +13,15 @@
   },
   "dependencies": {
     "bootstrap": "^5.2.3"
-  }
+  },
+  "name": "awesomebooks",
+  "description": "<a name=\"readme-top\"></a>",
+  "version": "1.0.0",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
 }
